requirements.txt
altgraph==0.17.2
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
filelock==3.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.28.5
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pefile==2021.9.3
Pillow==9.0.0
platformdirs==2.5.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.5
PyJWT==2.3.0
pyodbc @ file:///C:/Users/Olcia/Downloads/pyodbc-4.0.32-cp310-cp310-win_amd64.whl
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.31
twilio==7.7.1
urllib3==1.26.8
virtualenv==20.13.3
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
